Creating the task scheduler interface
The task scheduler interface can be used to schedule tasks to the event loop. They can be scheduled to execute immediately, at a specific time, recurringly after a specific time, the interface can reserve space for tasks in the event loop and this interface is able to return task queues. This is useful when you want to schedule a number of tasks at once.

Signed-off-by: HaMster21 <hamster.of.dev@gmail.com>