Creating a task queue interface
A task queue is used when the user wants tasks to be executed in a specified order instead of being executed in parallel by the ThreadPool. This allows for creating a dependency between tasks where the kernel assures that the tasks are executed in the given order, not starting execution of a task before the task before has finished running.

Signed-off-by: HaMster21 <hamster.of.dev@gmail.com>