Creating the recurring task handle interface
Implementors of this class will have control about when a recurring task is started and when the task should stop being executed. When the cancel method is called the task is not scheduled again. The Start method will execute the task and reschedule it after the specified time has passed.

Signed-off-by: HaMster21 <hamster.of.dev@gmail.com>