Restructuring with Visual Studio
The project file is now created by Visual Studio 2012 and defines a lot
more than the Mono one, Mono should be able to open it though. I thought
again about the folder structure of the project and decided that it is
sufficient to have a source folder that contains the specific project
files while the file VS created automatically are conatined in the
properties folder in the workspace root. The only files that are really
new are the settings class and the config mechanisms that Visual Studio
created automatically on my behalf. That should make it easier to read
from config files and even updating configuration while the application
is running (but that is not yet implemented).

Signed-off-by: HaMster21 <hamster.of.dev@gmail.com>