Moving project file and correcting build
This commit is introducing all dependencies to the project and makes it
build correctly. The only thing that may be confusing is the paths that
are expected by msbuild because the hint paths match my local settings
at the moment. I'm going to find out about that issue in the process of
building the project automatically on the build server. There is where I
have to resolve the dependencies automatically and this may influence
everybody else working with this project.

Signed-off-by: HaMster21 <hamster.of.dev@gmail.com>