Setting properties correctly and adapting project file
The Assembly info now reflects meaningful values.

The Settingsclass has been adapted to be used in the project. It can now
tell the kernel when the application settings are changes outside the
application during runtime.

The project file now expects a certain folder structure for building
correctly. The solution file is expected to be at ..\.. and the other
projects involved in the build are expected at
..\$(MSBuildProject.Name). That were the build process should find the
project files of the needed projects.