Writing the initial kernel
This first implementation of the kernel is not really doing anything.
When the main method inside the kernel class is run it creates a kernel
instance. This sets up logging and Settings, and is looking for the
config file. When found it sets the internal property for the
threadcount to this value, otherwise a default value is used.