adding kernel shutdown mechanism
When the kernel needs to be shut down there now is the shutdown method
to do so. It sets the internal state of the kernel to
isShuttingDown=true and every method that has to make sure the kernel is
running before getting to action will stop starting enything new. The
shutdown method will be the place to inform other components that the
server is going to halt.