Safer working with the kernel
The kernel class is now implementing the IDisposable interface, letting
us influence the way the kernel is treated by the garbage collector.

I also tweaked log4net to a certain extent and thought of the message
types in the kernel methods. Starting and shutting the server down is
now logged in info level. This is why there is a new info messages
ressource file.

The config file is now extended with the log4net configuration. This is
not currently running correct on my machine, but I'll find out why ;)